Adding makefile tbl2tssembl.mk and python script tssmerger.py
Running the makefile will take a tbl, fasta, sbt, and csv files
from $GOUPHOME/resources and outputs a merged embl file in the
working directory. The python script is executed by the makefile.

The default tbl file specified in the makefile needs to be updated
once Allyssa finishes her latest version of that file.

I've successfuly run the makefile in my local repo.