Updated default tbl input file to Allyssa's latest version.
Newest tbl file is $GROUPHOME/resources/H37Rv-hypotheticome.tbl
Resulting embl file is $GROUPHOME/resources/H37Rv-hypotheticome-tss.embl

I've also added some conditionals to the recipe for the temporary tbl
file created for tbl2asn, so they will always update the tbl header,
or add a header if none is present.

A few qualifiers in the tbl file still are not being transfered to
the output embl file. These are qualifiers labled "EC number" and
"ncnote" in the tbl file, instead of "EC_number" and "note".
I've placed the results into resources for now, but that will need to
be fixed.