Fixed TSS Strand Bug
Previous version of tssmerger.py added most TSSs from the Shell and
Cortes csv files to the wrong stran on the output embl. I determined
why, and fixed the code, and checked the new output. The resulting
TSS mRNA features in the output embl now have the same strand as their
equivalent TSS entries in the csv files. It is unfortunate that I
only just now caught the issue.