tssmerger.py now handles Antisense TSS correctly
According the Cortes et al paper where we got some of our
TSS data from, the TSS in the Antisense TSS file were labled
Antisense because they were found on the opposite strand of CDS
they were internal to. When I compared the Strand column of TSS
in that file with the CDS they were internal to, I found that
the strand column matched the strand of the CDS. Thus these
antisense TSS must be on the strand opposite what they are said
in the strand column (presumably that column refers to the strand
of the CDS?).

To address this, I modified tssmerger.py to flip the strands of
TSS it adds from the antisense csv file, and also lable them
'Antisense' in their note field.