Biopython Scanner.py is expecting six semicolons in the ID line of EMBL files
RATT outputs a deprecated version of EMBL files. The newest version that
biopython can parse requires six semicolons in the ID line. This was
fixed by adding a function that opens the file, fixes that line, and
appends the line to a list. All other lines are not affected and also
added to a list. The edited ID line and all other lines are then rewritten
to the original embl file.