Updated annotation directory object and added conditions in annomerge loop
Updated the annotation directory to a list of the references and output
in cases  where the user provided both. This required updating
downstream modules to account for a list of directory names including
creating a new list of all files in the target directories so that all
are combined.
Also added conditions to see if annomerge has already been run on that
genome and skipping annomerge if so.