Eggnog description is now present in the unannotated FASTA file
The Eggnog boolean description was missing from the FASTA header of the
unannoated FASTA file. This is resolved by adding a dictionary containing
all header IDs as keys and descriptions as values. The lookup is performed
when writing the reference proteome FASTA, which is later used to
identify unannotated sequences that have not been run through eggnog