Added a -f/--force argument
Added an argument that allows the user  to overwrite intermediate files
that were created  from a previous annotub run. This required adding
a condition that checks if the argument is not flagged and the
intermediate files exist, to print a message notifying the user that
the files were created but the -f/--force was not flagged and to do so
if the user wants annotub to run.