Fixed bug that was not inflating CDHIT clusters properly
Fixed a bug that was not inflating CDHIT clusters properly. In cases
where MCL clusters multiple CDHIT clusters, the data structure
is a list of multiple lists. Instead of just taking the first index,
the sublists are  iterated to flatten the list of lists. This list
is then sorted and all duplicate values removed.