Add docstrings/comments for temp clustering files
It was confusing where all the files in the temporary clustering/
folder come from and how they're related. I added more explanation
to the docstrings of the functions that produce these files. I also
added comments where these functions are called to specify what files
are being used as input or writtten as output.