Check if unannotated_fasta is empty before BLAST
If top_hit is None, set assign_mtb to True right away and then check
if unannotated_fasta is empty before trying to run BLAST. If this
fasta file is empty BLAST will throw an error. This came up because
if the reference genome doesn't have MTB#### genes then
unannotated_seqs.fasta is empty, leading to the BLAST error.