Move gbk_to_gff converter call. closes #15
Because this converter call was within the run.eggnog_mapper call,
sometimes it wouldn't happen because if eggnog_seqs.fasta doesn't
exist then run.eggnog_mapper isn't called. By pulling it out separately
this ensures the conversion step happens and the GFFs are updated
regardless of whether eggnog is run or not.