Allow for a read-only references directory
Conversions to embl and gff of the genbank files in the user-provided
directory now all happen in the temporary references folder,
which now also contains symlinks to the original gbk files so that
this directory can be used as a single source.

mitigates #8 since a large part of the issue involved refdir joins
that happen in this function, which have been fixed.