Add coverage flag, change identity flag, use int
Note that the flag for identity threshold is now -i not -c.

Also, these thresholds are mostly used for blastp which takes an integer
and not a float (which CDHIT takes). So instead of converting to a
float everywhere I decided to change the parameters to take an
integer, then convert to float within the function that runs CDHIT.