Add inference tag labeling RATT as the source of its annotations
Rather than log a file saying which source was responsible for every
gene in the final annotation, we can just have the inference tag label
the source of the annotation, as described by a /inference tag. See
https://www.insdc.org/files/feature_table.html#7.3.1

Although RATT should probably be adding this in itself, we can do it
in the meantime.