correct_start_coords_prokka: use deepcopy before making start shifts
There is no bug as far as final results due to this because
modifications made were undone if they turned out to be rejected by
the criteria, but the loggging was affected since it depended on
comparing the original to the modified. Using deepcopy avoids this
problem and also avoids the need to deal with changes to the original copy.