extractor.py: generalize usage and standardize names
Also, SeqIO.write can use both a file handle or a file name string, so
we don't need to wrap invocations with `with open(` after all.
But to take advantage of using a string filename argument, I needed
to write the CDSs in a single statement rather than in a loop so that
the file is not overwritten with each iteration. SeqIO.write can use
a list and doesn't need to be wrapped in a loop anyway.