BLAST: make blastp() operate on SeqRecords directly
And so we no longer need the function to loop through the SeqRecords
and pull out seqs and ids. This also has the added benefit of allowing
us to selectively care about the sequence IDs, because we will not be
needing them when we start using this function from annomerge.