Properly calculate blast query/subject coverage
Gaps in the alignment were getting counted as part of the aligned
sequence, so we were sometimes seeing qcov/scov > 100%. This commit
will close #27 once refactoring is complete and we have just one
blastp function in use everywhere.

I wanted to make this change after the refactoring, but the
refactoring also required making a dictionary of the blast results,
which I didn't want to do with the incorrect coverages.