BLAST refactor - part 3: remove all blast_feature_sequence_to_ref()
blast_feature_sequence_to_ref had a check for whether the reference
gene had a translation and, if it didn't, the blast comparison was
skipped and the annotation in question automatically accepted. This
situation appears to only be the case for pseudogenes, so I added a
step in extractor to generate translations for pseudogenes so that we
can have something to compare to.