BLAST: properly handle retention of top hit per subject
The summarize() function was retaining the last displayed hit per
subject. For cases where there's only one hit per subject, this
doesn't matter, but for other cases, this means that the
lowest-scoring hit was the one retained since blast displays results
sorted in descending order by e-value. This is easily fixed by going
through the blast results in reverse order so the last thing we see,
which is what we keep, is the top hit for that subject.

This also means that the dummy 0%/0%/0% hit needed to be added to the
end of the blast output to play along.