add extractor.subset_fasta and use it to generate a separate reference fasta
various stages of parseClustering blast against a "reference_fasta"
and then the blast code filters out hits matching unannoated
sequences. Since we want to use BLAST.blastp and shouldn't have to
exclude subject sequences after the fact, we should avoid blasting to
them to begin with. The find_unannotated_seqs() function here is
generalized, moved to extractor.py, and used in parseClustering to
generate an actual reference fasta. Then this is passed along to all
the functions that expect it.