BLAST refactor - part 4 of 4: remove extra blastp functions in parseClustering
I had to add a top_hit() function to BLAST.py as a thin wrapper around
max(), but parseClustering's identify_top_hit was preserved for the
way it collected the results. It could potentially also be absorbed into
BLAST.py if we end up needing something similar elsewhere in the pipeline.