annomerge: split discontinuous ratt annotations into two features
The gff conversion of a gbk entry with joins is not meaningful,
and causes some problems, as the entire sequence gets labeled
"biological region" and two other basically empty CDS records are
created. Since we create a parent 'gene' feature for these later,
splitting them now isn't a big deal.