parseclustering: better matching of reference and unnamed genes
There was still a stray "Rv" here that was used for matching (TB)
reference genes that's now switched to using designator.is_reference.
Also, many checks in this part of the program intending to find prokka
ltags search for strings that start with "L", since "L_" wouldn't
match things like "L2_". This is somewhat dangerous since there are
reference genes that start with "L". So we now use
designator.is_raw_ltag for these checks instead.