check_inclusion_criteria: call validation function once before conditional
validate_prokka_feature_annotation gets called with exactly the same
parameters in every branch of the conditional block. I believe the
original author did this because this function modifies its
input (prokka_annotation) and that variable is used for checking some
of the conditions. The function probably shouldn't be doing that, but
we can at least avoid that issue here by passing it a deepcopy. So now
we only call validate_prokka_feature_annotation once in this function.