Only run Prokka once
Run it without the reference proteome. Next to implement is the
reference check of prokka annotations within annomerge. After that,
all reference name assignments can be accepted as they are since we'll
have determined them using our criteria from the start. The stage is
set here to accept all names as they are.