fileManager: make sure all reference CDS annotations have gene names
Many parts of the code dealt with checking whether a gene qualifier
existed and falling back to the locus tag if it doesn't. And I think
other parts of the code should be doing that and don't. We can just
make the whole thing easy and ensure our starting reference
annotations always have the gene qualifier.