annomerge: apply liftover if ab initio is found to overlap in-frame with ratt transferred-gene
The reference name was already getting
assigned for the cases of in-frame RATT-transfer overlaps,
with the decision to include the ab initio version falling to
check_inclusion_criteria. Part of that check involves comparing the
blast hits to the reference, which we would not have saved if the direct
comparison of the ab initio annotation to the reference didn't meet
the thresholds. So now we save all blast hits in case we need them later.