check ref_match rather than blast_stats for validating RATT
This only matters if `--filter-ratt` is true, but, since our BLAST
function now always returns the blast summary, whether there's a
passing match or not, and because we always have a null hit in order
to pass on 0-thresholds (i.e. not filtering RATT), this will always be
true and always result in a `valid` call. ref_match being defined is
the proper thing to check.