annomerge: add logging for parallel blasting
Also get rid of the redundant log message in annomerge.run() about the
number of RATT CDSs. It's already logged in the RATT validation
function (with the nuance of not counting CDSs we broke in two pieces
that would be annoying to have to propagate out of the function to
make the other log message I just removed give a consistent count).