Force order of contigs processing as in fasta file
We can't rely on the listdir sorting order of the ratt annotation
files to properly order the contigs. We need to be able to compare the
RATT and Prokka annotations in the same order, and the Prokka genbank
file naturally follows the order of the input fasta while RATT has
each contig's results in a separate file.