annomerge: Don't run check_inclusion_criteria if one of the conflicting parties has been abolished
The test case here is 4-0041's Rv1718 and L_02383. The Prokka
annotation is exactly the same as the reference H37Rv, while RATT's
annotation transfer extended the start position 6 bases, so
check_inclusion_critera() rejects the RATT annotation. Apparently this
gene overlaps another Prokka annotation, so when the code tries to
look up the RATT annotation again, it fails because it's been
deleted. So we now check to see whether the annotation is still
present and, if it isn't, skip the comparison.