reduce input requirements of check_inclusion_criteria
This function doesn't need to handle deleting the ratt feature from
the dictionary and populating the ratt rejects list. I also drop the
need for passing ratt_gene_location since this information is included
in the ratt_feature itself. Soon, we'll also be able to drop
ratt_rejects as a global variable.