test_annomerge: fix sequence context check to draw from reference
The promoter check in check_inclusion_criteria is supposed to check
the sequence context of the start position in the isolate's ratt
annotation vs that of the reference. I was incorrectly giving it
the sample sequence and ratt annotation, which is of course always a
perfect hit :facepalm:. The test now accurately calls the function the
way hybran does.