annomerge: don't send mutations from the same source to the thunderdome
The final feature annotation verification looks for CDSs with the same
start or stop position and assumes one is from Prokka and the other is
from RATT. In reality, some of these cases are conjoined genes that
RATT has annotated and both of those annotations should be kept.

This change also combines an if/elif into a single conditional because
both branches had the same body.