annomerge: precisely determine name of RATT output files
If the contig name has a '.' in it, we don't correctly figure out the
contig name from the file name (since '.' also separates the sample name and
the contig name and...what if the sample name were to have a '.' in it as
well). Since we know that RATT output files are in the form of
  sample.contig.final.embl
and we already know the sample and contig names, we know exactly what
the file name should be. The only caveat is that RATT replaces '|' in the
sample and contig names with underscores, so we do that here too.