annomerge: replace pick_best_hit with check_inclusion_criteria
The two functions are almost the same. Although pick_best_hit
guarantees that only one annotation is selected, it is not necessarily
always the right call. pick_best_hit() also has mistakes in the
(duplicate) code to the check the upstream sequence context since it
was checking for strand '+' rather than 1. The only specific thing
that pick_best_hit did was a blastn of the RATT feature to the
corresponding reference feature.