annomerge: prevent resetting of gene name for annotations with domain inference
This code is leftover from when we kept certain Prokka name assignments,
but we didn't want those that were made based on domain matches.
It's causing trouble now since (1) we now throw out all Prokka name
assignments early on, so name assignments at this point came from our own
reference matching and should be kept and (2), if the gene matched as a
`pseudo` to the reference, it cannot recover its name in the clustering
step since known pseudo sequences are not used for clustering.
This contributes to #43.