annomerge: generate reference cds fasta files before processing sample annotations
The reference contigs don't necessarily correspond one-to-one with the
sample contigs, so their information should be collected separately from
the sample contig annotation loop. In this case, the information is the
CDS nucleotide sequences for the genes with start positions incorrectly called by prodigal, as well as upstream context sequences for all reference CDSs (since they are expected for RATT/ab initio comparisons in check_inclusion_criteria).