Add test case for inclusion criteria for when RATT has better alignment coverage
I noticed while testing the development version that spurious
duplicates are arising that did not used to occur in version 1.5.2.
Removing the redundant Prokka/RATT comparison code seems to be
responsible for this (not yet released!)
regression. check_inclusion_criteria is currently not handling this
case properly, as shown by this test case's failure.