annomerge check_inclusion_criteria: start with prom_mutation=True
If we have no hit at all, we don't get the opportunity to set
prom_mutation=True, so the better path is to start with it set to True
and then change it to False if we find a perfect hit.
This corrects the handling of the Rv1453 test case I just added,
although that should not be returning 0 hits from blastn. That's a
separate problem, but this change remains sensible and valid.