annomerge: check _all_ RATT/prokka in-frame conflicts for identity
Before, we only did this if there was a single conflicting pair, but
this is wrong: what we actually needed to look for was a mutually
exclusive pair. That would be Prokka and RATT annotations that are
in-frame and very likely represent the same gene. If the ab initio
annotation mistakenly didn't get a name assigned (due to thresholds),
we're likely to recover the assignment by comparing to the
RATT-transferred instance of the annotation. And in this case,
check_inclusion_criteria can make sure only one ends up in the final result.