annomerge: don't blast the ab initio genes that immediately get thrown away
Since we no longer use the gene name to assess redundancy of ab initio
annotations with those from RATT, we can run
find_inframe_overlaps *before* doing the ab initio blasts. In doing
so, we'll identify all the identical CDSs that we're going to throw
out, so we only have to run blast on the remaining ones. This saves
close several minutes even when using 12 cores.