find_inframe_overlaps: use end position to determine whether to break loop
Changing this from abinit_start to abinit_end didn't make a difference
in the results for 1-0006, but this looks to be the more appropriate
reason for determining that no further genes will overlap. Whereas if
a RATT annotation had a later start position than the ab initio
annotation, the original criterion would have overlooked it for
comparison. I suppose I should add a test case along these lines.