annomerge: fix overlap_inframe to work with ORFs containing internal stops
pseudogenes can go past their internal stop codons, so it's not
sufficient to check for in-frame overlaps by only comparing strands
and stop positions. Changes here now do a comprehensive check from
both ends of the gene. Since pseudogenes could have lengths that are
not multiples of 3, either its start-defined reading frame or
stop-defined reading frame can be used to determine if we have an
in-frame overlap with something else.