annomerge: Add have_same_stop() and use it in identify_conjoined_genes()
This new function behaves identically to how overlap_inframe() used
to. The latter function is now to general to be adequate for the
purposes of identify_conjoined_genes since it will flag cases like
espH and eccA1 in 1-0006, where eccA1 has an early internal stop, so its
first few bases are in-frame with the upstream gene espH, but eccA1's
actual stop codon is in a different reading frame and much later. So
if we go with overlap_inframe, these two are considered conjoined when
that is not the case.