annomerge: Remove the reference-prodigal step from the pipeline
We're doing start coordinate correction differently in that we no
longer rely on offsets to whatever Prodigal called on the reference
genome. Applying those same offsets to a sample genome can fail for
situations like in-frame insertions, or in case Prodigal doesn't call
the same corresponding start that it would have called in the
reference genome for other reasons.

This removes our direct dependency on prodigal (though we will still
have it because Prokka depends on it) and it also facilitates our
future switchover to alternative ab initio engines that might not use
Prodigal.