annomerge: write ab initio coord corrections to the prokka folder
It precedes the actual annotation merging and not all of these
corrected annotations will end up in the final result. This part is
really just a postprocessing of the ab initio annotation, so should
be written there. This will also remain consistent with the way it was
done before we introduced coord_check, so no documentation updates
regarding output file name/path will be necessary.