isolate_valid_ratt: remove check for existence of Location
We've never actually encountered a RATT annotation that was missing
the Location, but someone felt the need to add this code in. The fact
that it was far down in the else/if chain--not to mention after other
else/ifs that were testing the very Location field-- shows that it's
not needed at all.