Genes marked for correction, where the alignment was longer than both the reference and the feature were getting incorrect coordinates.
This can only occur if the aligner puts gaps in both sequences. It's more clear if a gene position is being modified because of the distinction between 'feature' and 'pad_feature'. Also, included a check for an alignment score if a gene position is modified. Genes marked for correction shouldn't have a worse alignment after adding up/downstream context.