Added "corrected_feature" to keep track of corrected coordinates.
If the second alignment score was only worse because of excess padding - we would run a third alignment for verification. However, this third alignment was incorrectly using the unmodified feature. Now we are using the corrected feature in the third alignment. Fixed for loop to iterate more than just once.