Add a specific amount of upstream context rather than simply the difference in gene lengths.
There is a scenario where the difference in base pairs between two genes in an alignment does not equate to the correct amount of context needed to correct coordinates. In Rv2280 (isolates 1-0006 vs H37Rv) the gene lengths are very similar, however there is little overlap between the genes in the alignment. The difference in gene lengths is not enough to correct the coordinates and find good starts/stops. Now instead, we calculate a specific amount of context needed (plus a little extra (4%)) to find the starts and stops separately on each side of the alignment.