Fixed ambiguous variable name and replaced arbitrary percentage when adding context to a gene.
Before, we would use the difference in length between the reference gene and the feature and add 4-5% of this length as an extra buffer for the alignment. Now, the length of the interval used to determine found_low/found_high for each gene will instead be used as this buffer when adding context (padding) to a gene.