Reference correspoding starts were not getting corrected if the end of the gene required downstream context.
If we are adding up/downstream context where one start or stop already corresponds to the reference, we should
use the original found_low/high variables to determine if correction is also necessary. This issue came up in
PE10 in pseudoscan unit test. The original alignment and the padded alignment both found a reference
corresponding start and had the same alignment score, so the feature start was never corrected when it actually should have been.