Update name and function of BLAST.reference_match return value.
Previously we were only looking for low subject coverage and passing query coverage to determine
if a gene was 'pseudo'. Now we also look for the opposite (low query coverage, passing subject coverage).
The variable name was changed to low_covg because the coverage information doesn't necessarily mean
a gene should be marked pseudo.