pseudoscan: revise criteria regarding keeping coord corrections
see comments in code. The case of a marginal coordinate correction
that restores reference correspondence when the original coordinates
had a good enough blastp hit was not being handled properly.
The length filtering that was in place before was only intended to
reduce the set of genes we checked, not be a part of pseudoscan
itself. In any case, we're just going to check all the genes instead
of putting that filter back in isolate_valid_ratt_annotation().