WIP: Create more descriptive notes for specific types of pseudo genes
We have enough information to declare if a gene has an in-frame or out-of-frame indel.
We can also detect SNPs and alternative stop codons. Adding this information can be
really helpful in the final annotation because it can provide extra context for genes that are
unexpectedly marked pseudo/non-pseudo. Also, we've added a temporary 'code' to quickly
view the basic characteristics of a pseudo/non-pseudo gene (ie. divisible by three,
internal stops, good coords etc) which may get added to a separate log file in the future.