parseClustering: don't make pseudo calls based on clustering matches
pseudogenes identified in annomerge are excluded from clustering. We
now require complete (full alignment coverage) matches to assign a
name at this point so that we can be reasonably assured that our
matches are non-pseudo. Ideally, we would allow low-coverage matches
and run pseudoscan if we get a hit, but we don't have easy access to
all the data we need and we'd also need to repeat some steps in
annomerge, like process_split_genes. It needs more thought and
attention than we have time for at the moment.