Catch corner cases for inverted 'join' genes.
In this test case, the original alignment was completely off (found_low/found_high = False). After adding
up/downstream context, coord_check found a reference corresponding start past the end of the original gene,
but failed to find the reference corresponding stop. Coord_check tries to correct the start, while leaving
the stop alone and encounters the "End Location must be greater that start location" error.

Now if this happens, we won't correct any coordinates and we will leave the gene as it was.