Insure that upstream genes with nonstop mutations are identified correctly.
Sometimes the wrong gene was being tagged pseudo because we weren't actually checking
the locations of a feature to determine if it was up/downstream of the previous gene.
Partially addresses the problems in issue #51 because idsB becomes rightfully non-pseudo.
But now there's an extra copy of lytB1 that needs to be eliminated.