fusionfisher: match same-location features by gene rather than ltag
The original rationale was that it was for replacing the functionality
of the old `remove_duplicate_cdss` function that was part of RATT
postprocessing, but this function now is also used in
`check_inclusion_criteria`. In this case, locus tags will never match
between RATT and ab initio annotations and we should match against
gene names instead.