WIP: add data for inclusion_criteria test case
See https://gitlab.com/LPCDRP/hybran/-/merge_requests/32#note_1372590148
This is a case where one gene is pseudo and the other isn't. But the
gene that is pseudo is because its reference annotation is pseudo. So
this is an interesting case where we have a better match to the intact
second copy of the gene than we do to the copy that is syntenically matching.