annomerge: remove step where we blast unnamed abinit genes to RATT genes
This was originally intended to try to salvage a name assignment for
the ab initio genes. If this happened, they entered check_inclusion_criteria
as a conflicting annotation pair, which used to be helpful when we
had more limited postprocessing and we could choose which annotation
was better delineated. At this point, the annotations are most likely
to be identical or overlapping enough that blasting them together
is likely to result in a complete hit.

TL;DR if the ab initio gene is unnamed by this point, it gets what's
coming to it.