Ensure that we never correct features with invalid coordinates.
In some rare instances, coord_check will try to correct a reference corresponding
start past the end of a feature, or attempt to correct a reference corresponding
stop before the start of a feature. In both scenarios, a ValueError will arise
when creating the newly 'corrected' feature and crash the program. This problem occurs when
the original alignment is really far off and we find a downstream start/upstream stop by
chance from the addition of extra padding. We've implemented a fix for this in the past,
however, it was only being considered under specific circumstances. Now, this exception
is considered for every potential coordinate correction. Also, added test case with feature
that caused this issue.