stopseeker: only return features with valid in-frame stops.
Fixed a bug where we were extending a sequence by the wrong amount of base pairs. We
were checking the length of the translated sequence and multiplying by three, when we needed to just
look at the length of the original feature. Also implemented an additional coord_check and
has_broken_stop call to update the feature's 'coord_status' and 'stop_note'.