Make use of all available references throughout the pipeline
We had a primary reference that was the only one being used as a blast
reference for ab initio annotations. Multiple references were only
really being used by RATT. We now use the unique ref cds fasta file
from onegene as the blast reference. Expected fixes are also made for
the eggNOG stage of the pipeline, but that was already in disrepair
and not really well-supported.

still to do:
- annomerge's ref_annotations dictionary should have another level
  where the gene annotations can be looked up by reference genome.
- extractor.gene_dict should similarly make a nested dictionary