WIP: annomerge: build in support for comparing to multiple references
- Make ref_annotation a flat dictionary keyed like ref_id@@@ref_gene_name to
allow distinguishing between alternative but not identical copies of
the same gene from different references.
- Monkeypatch in an attribute to SeqFeature: .source, which will hold
the ref_id of the reference from which the annotation was lifted
over. So far, this is only for ab initio annotations. I still need to
work it out for the RATT annotations.
- Adjust ref_fuse for the possibility that the components of a fusion
gene may not draw their best-matching reference from the same
genome. This requires being able to feature.extract different parts of
sequences from different contigs/chromosomes. feature.extract doesn't
support this, so I created extractor.get_seq() to handle this.
  - as part of this change, we no longer use the monkeypatched
  `hybranref` attribute to SeqFeature. Instead, we always use the
  feature.location.parts[i].ref attribute. location.parts always
  exists, even for SimpleLocations, so this is a reliable way of
  accessing the sequence contig corresponding to the coordinates. This
  function is currently used only with reference features. We'll
  likely be subclassing CompoundLocation/SimpleLocation in the near future.

Remaining to do:
- update test cases
- Set feature.source for RATT-transferred annotations