Add a tracer note to our temporary references so we can see where RATT is pulling from
RATT will transfer annotations from the marked references (which are
in the temporary directory, so we don't tamper with the user's copy),
then we read that in to set the source attribute and discard the
note. We also discard the note when we load in the reference
annotations so that they don't get propagated to the ab initio annotations.