Streamline and fully parallelize RATT feature validation
pseudoscan was not previously included in the parallelization and it's
a slow step. Changes here also allow for skipping blast reference
matching altogether when enforcing blast thresholds on RATT annotation is not
desired, which should save a good amount of time in that case. This
commit also makes sure to incorporate RATT-transferred non-CDS
elements that were lost following the initial refactor of the
postprocessing steps out of annomerge (WIP commit).

**Parallelizing of pseudoscan causes the coordinate correction report
data to be lost. We need to work out another way of collecting that.