Don't initialize og and corr properties from the constructor header
There seems to be a problem with calling a constructor as a default argument
(when using multiprocessing?). The coordinate correction log was showing
the same results for feature.og.location and feature.corr.location as other
features, but the other attributes seemed to be intact and correct.
Initializing the og and corr attributes in AutarkicSeqFeature this way
fixes the issue.