Intercept parser errors and exclude offending features
various RATT issues due to rolling their own EMBL writer result in
malformed location strings (see
https://github.com/ThomasDOtto/ratt/issues/12 and
https://github.com/ThomasDOtto/ratt/issues/18 ) that normally crash
the Biopython parser. Here, I subclass the parser and make it a little
more robust to parsing errors. We either lose the offending feature
entirely during parsing, or bring it in with its location set to
None. We then check for the latter when validating the RATT features.