WIP: Modification of delayed stop criteria in coord_check
All features should have a delayed stop attribute. All delayed stops should be 'good_stop = False'.
If a feature sequence extends beyond the last b.p. alignment to the reference,
it should always be considered a delayed stop. If a spurious alignment from the reference
sequence matches with the end of a feature, it should be considered a delayed stop if
the alignment block is less than our calculated interval of reference-correspondence (ex. 'Rv0325')

Also, coord_check shouldn't be responsible for determining if a feature is accepted/rejected.
Allow that decision to be made in pseudoscan.
be responsible for determining if corrections are accepted/rejected.