Repair fissionfuser unit test following AutarkicSeqFeature addition
This test wasn't fixed along with the others because it would have
required some more extensive changes when we wanted pseudoscan to
update the blast hits dictionary and fissionfuser had to require that
as an argument. We've done away with that recently, so this test update
now is not so involved.