fissionfuser: abort fragment fusion if one turns out to be non-pseudo
Despite appearing to be complementary fragments, ORFs can still be
intact copies and therefore called non-pseudo if they have a >=95%
blastp hit by sequence identity and alignment coverage. In this case,
they actually should be treated as independent ORFs, but something
interesting is happening regardless, so we log something to the
screen.

closes #66