fissionfuser: separate confirmation and execution steps
allows us to make one single check for 'overlap_inframe', which should
automatically pass. The changes made for issue #66 didn't explicitly
rule that out, but it was unlikely that a non-pseudo would overlap
in-frame since we have only seen that overlap due to encompassing an
internal stop, but better safe.