log_feature_fates: fix handling of empty fields
For the case of malformed RATT features, we got a KeyError on
locus_tag. I now replace this in validate() with the feature.id and
then have log_feature_fate systematically replace None with '.'. We
also now do this in one place in the funcion, which will make changing
our empty field character easier should we need to in the future.