test_find_inframe_overlaps: revise expected output for double_overlap
Mark both RATT annotations as conflicting rather than just one. This
makes sense since they both are in-frame overlaps (one has same start
and the other has same stop). It also doesn't change the final results
of the run after re-testing, so perhaps this was just an oversight
when creating the test initially. Especially since the
having-same-start criterion for overlap_inframe() (which the tested
function depends on) is relatively new.