Modify coord_check scoring parameters to help identify delayed ends.
We've increased the internal_extend_gap penalty to further dissuade
the aligner's tendency to align delayed stops at the end of the query
sequence. Introduced the 'relaxed_found_high' variable to help with
the assignment of the delayed end attribute. Now we can identify
delayed ends where frameshifts/mismatches occur at the end of a sequence
(in the final alignment interval).