ensure final feature verification still happens when one source is empty
When RATT or Prokka find no annotations for a contig, we were
using whatever the other source had and continued the loop. There is
some important postprocessing that happens at the end of the
loop (like setting locus tags as gene names where missing and removing
translations for pseudos). So what we do now is add back in the `else`
clause that we used to have so that it only runs when both annotation
sources have data, and the final postprocessing happens in all cases.