WIP: Modification of 'relaxed_found_high' criteria
There are a lot of imperfect alignments, and it's difficult to
distinguish between legitamate structural variation vs noise from
the aligner. We attempt to mediate this by introducing the 'relaxed_found_high'
variable which is meant to be a less strict version of 'found_high'.
Specifically we use this variable for the identification of delayed stops.
Previously we weren't taking into account the possibility of gaps appearing
on both sides of the alignment. This became a problem, specifically when
we began digging deeper into the alignments that violated/circumvented the new
'relaxed' criteria. Here I created a function to help retrieve the alignment
sequences with gaps included for the purpose of solidifying our 'relaxed' criteria.